requirements.txt
streamlit==1.52.2
pandas==2.3.3
pytest==9.0.2
requests==2.32.5
pydantic==2.12.5
mysql-connector-python==9.5.0
matplotlib==3.10.7
uvicorn==0.38.0
plotly==6.5.2
fastapi==0.125.0